requirements.txt
pandas
torch
transformers
regex
praw
kafka-python
cassandra-driver
pyspark==3.5.0
python-dotenv